+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.7.5 - Unreleased
 
+### Added
+* `load_sample` now takes an `annot_ids` argument to skip the spatial lookup
+  and use something precomputed.
+
 
 ## Version 0.7.4 - Released 2023-03-05